Hosmer-Lemeshow Goodness-of-Fit Test

Performs the Hosmer-Lemeshow goodness-of-fit test for a fitted logistic regression model. Observations are sorted by fitted probabilities, divided into g groups, and a chi-squared statistic is computed comparing observed and expected event counts within each group.

Usage

hoslem_test(model, g = 10)

Arguments

model

A fitted glm object with family = binomial.

g

Integer. Number of groups (default 10).

Value

A list of class "hoslem_test" containing:

statistic

The Hosmer-Lemeshow chi-squared test statistic.

df

Degrees of freedom (g - 2).

p_value

P-value from the chi-squared distribution.

table

A data frame with observed and expected counts per group.

g

Number of groups used.

References

Hosmer, D.W., Lemeshow, S., and Sturdivant, R.X. (2013). Applied Logistic Regression, 3rd Edition. John Wiley & Sons.

Examples

m <- glm(am ~ mpg + wt, data = mtcars, family = binomial)
hoslem_test(m)
#> Hosmer-Lemeshow Goodness-of-Fit Test
#> ------------------------------------
#> Groups: 10
#> Chi-sq: 11.6248  df: 8  p-value: 0.1687
#> 
#>  group n observed_1  expected_1 observed_0 expected_0
#>      1 4          0 0.003917959          4 3.99608204
#>      2 3          0 0.067663822          3 2.93233618
#>      3 3          0 0.203667175          3 2.79633282
#>      4 3          0 0.321092385          3 2.67890762
#>      5 3          1 0.436171580          2 2.56382842
#>      6 3          0 0.571211040          3 2.42878896
#>      7 3          3 2.010473871          0 0.98952613
#>      8 3          3 2.535787992          0 0.46421201
#>      9 3          2 2.888464754          1 0.11153525
#>     10 4          4 3.961549421          0 0.03845058