Overdispersion Test

Tests for overdispersion in a GLM by comparing the sum of squared Pearson residuals to its expected value under the assumed model. Most useful for Poisson and negative binomial models, but can be applied to any GLM where the dispersion parameter is assumed to be 1.

Usage

overdispersion_test(model)

Arguments

model

A fitted glm object.

Value

A list of class "overdispersion_test" containing:

dispersion

Estimated dispersion (sum of squared Pearson residuals divided by residual degrees of freedom).

statistic

The Pearson chi-squared statistic.

df

Residual degrees of freedom.

p_value

P-value from the chi-squared distribution (upper tail).

family

The model family name.

Details

The test computes sum(pearson_resid^2) and compares it to a chi-squared distribution with n - p degrees of freedom. A dispersion estimate substantially greater than 1 suggests overdispersion. A significant p-value indicates the variance is larger than expected under the assumed distribution.

Examples

counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
m <- glm(counts ~ outcome + treatment, family = poisson)
overdispersion_test(m)
#> Overdispersion Test (Pearson Chi-Squared)
#> ------------------------------------------
#> Family: poisson
#> Dispersion estimate: 1.2933
#> Pearson Chi-sq: 5.1732  df: 4  p-value: 0.2700
#> => Evidence of overdispersion (dispersion > 1)
#>    Consider: quasipoisson, negative binomial, or robust standard errors